% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-adm.R
\docType{data}
\name{adm_0_sf}
\alias{adm_0_sf}
\alias{adm_1_sf}
\alias{adm_2_sf}
\alias{adm_region_sf}
\title{Mongolia administrative level 0-2 boundary polygons, lines and gazeteer.}
\format{An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 1 rows and 4 columns.}
\source{
\href{https://data.humdata.org/dataset/mongolia-administrative-boundaries-polygon-polyline}{UN HDX}
}
\usage{
adm_0_sf

adm_1_sf

adm_2_sf

adm_region_sf
}
\description{
Mongolia administrative level 0 (country), 1 (aimag), 2 (soum), and regional boundary polygons and lines, and administrative gazeteer.
}
\examples{
head(adm_0_sf)

head(adm_1_sf)

head(adm_2_sf)

head(adm_region_sf)

}
\author{
Brendan Knapp \email{brendan.g.knapp@nps.edu}
}
\keyword{datasets}
