% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-alagac.R
\docType{data}
\name{alagac_roads_sf}
\alias{alagac_roads_sf}
\alias{alagac_railroads_sf}
\alias{alagac_aimag_centers_sf}
\alias{alagac_aimag_boundaries_sf}
\alias{alagac_soum_centers_sf}
\alias{alagac_soum_boundaries_sf}
\alias{alagac_national_boundary_sf}
\alias{alagac_major_lakes_sf}
\alias{alagac_major_rivers_sf}
\title{Data from the Administration of Land Affairs, Geodesy, and Cartography of Mongolia (ALAGaC)}
\format{An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 2172 rows and 6 columns.}
\source{
\href{https://marine.rutgers.edu/~cfree/gis-data/mongolia-gis-data/}{marine.rutgers.edu}
}
\usage{
alagac_roads_sf

alagac_railroads_sf

alagac_aimag_centers_sf

alagac_aimag_boundaries_sf

alagac_soum_centers_sf

alagac_soum_boundaries_sf

alagac_national_boundary_sf

alagac_major_lakes_sf

alagac_major_rivers_sf
}
\description{
Data from the Administration of Land Affairs, Geodesy, and Cartography of Mongolia (ALAGaC)
}
\examples{
library(ggplot2)

ggplot() +
 geom_sf(aes(color = "alagac_soum_boundaries_sf"),
         fill = "transparent",
         data = alagac_soum_boundaries_sf) +
 geom_sf(aes(color = "alagac_aimag_boundaries_sf"), 
         fill = "transparent",
         data = alagac_aimag_boundaries_sf) +
 geom_sf(aes(color = "alagac_national_boundary_sf"), 
         fill = "transparent",
         data = alagac_national_boundary_sf) +
 labs(title = "Administrative Boundaries") +
 NULL

ggplot() +
 geom_sf(data = alagac_national_boundary_sf) +
 geom_sf(aes(color = "alagac_aimag_centers_sf"), 
         fill = "transparent",
         show.legend = "point",
         data = alagac_aimag_centers_sf) +
 geom_sf(aes(color = "alagac_soum_centers_sf"), 
         fill = "transparent",
         show.legend = "point",
         data = alagac_soum_centers_sf) +
 labs(title = "Administrative Centers") +
 NULL

ggplot() +
 geom_sf(data = alagac_national_boundary_sf) +
 geom_sf(aes(color = "alagac_major_lakes_sf"), 
         data = alagac_major_lakes_sf) +
 geom_sf(aes(color = "alagac_major_rivers_sf"), 
         data = alagac_major_rivers_sf) +
 labs(title = "Major Bodies of Water") +
 NULL

ggplot() +
 geom_sf(data = alagac_national_boundary_sf) +
 geom_sf(aes(color = "alagac_roads_sf"), 
         data = alagac_roads_sf) +
 geom_sf(aes(color = "alagac_railroads_sf"), 
         data = alagac_railroads_sf) +
 labs(title = "Transportation") +
 NULL

}
\author{
Brendan Knapp \email{brendan.g.knapp@nps.edu}
}
\keyword{datasets}
