\name{BlueSky-package}
\alias{BlueSky-package}
\alias{BlueSky}
\docType{package}
\title{
Open various file formats, run various inbuilt analysis on them and get output in text, html or LaTex format.
}
\description{
This package can be used to open various file formats(RDATA, CSV, DAT, XLS, XLSX, SAV, DTA, SAS7BDAT) and run various inbuilt analysis on them producing output in text, html or LaTex format.
}
\details{
\tabular{ll}{
Package: \tab BlueSky\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2021-10-15\cr
License: \tab GPL (version 3 or later)\cr
}
Package can be used with BlueSky Statistics application or any other R environment

}
\author{
BlueSky Statistics.

Maintainer: contact us <support@blueskystatistics.com>

}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ 
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
