% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapiro wilks test.R
\name{BSky_Shapiro_Wilk_normality_test}
\alias{BSky_Shapiro_Wilk_normality_test}
\title{Shapiro-Wilk Normality Test}
\usage{
BSky_Shapiro_Wilk_normality_test(data = NULL, vars = NULL, dataset = NULL)
}
\arguments{
\item{x:}{a numeric vector of data values. Missing values are allowed, but the number of non-missing values must be between 3 and 5000.}
}
\value{
A list with class "htest" containing the following components:
statistic: the value of the Shapiro-Wilk statistic.
p.value: an approximate p-value for the test. This is said in Royston (1995) to be adequate for p.value < 0.1.
method: the character string "Shapiro-Wilk normality test".
data.name: a character string giving the name(s) of the data.
}
\description{
Performs the Shapiro-Wilk test of normality.
}
