% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UA_UI_Functions.R
\name{BSkycloseDataset}
\alias{BSkycloseDataset}
\title{BSkycloseDataset closes a dataset, removing it from the R.}
\usage{
BSkycloseDataset(dataSetNameOrIndex)
}
\arguments{
\item{dataSetNameOrIndex}{is the name of the dataset that is to be closed.}
}
\description{
BSkycloseDataset closes a dataset, removing it from the R.
}
\examples{
df <-data.frame(A=c(1,2,3), B=c(4,5,6), C=c(6,7,8))
BSkycloseDataset(df)
}

