% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.r
\name{BSkyConfusionMatrix}
\alias{BSkyConfusionMatrix}
\title{Confusion matrix}
\usage{
BSkyConfusionMatrix(
  modelname,
  showConfusionMatrix = FALSE,
  predictions,
  datasetname
)
}
\arguments{
\item{modelname}{a model object for which confusion matrix is desired.}

\item{predictions}{an object that is returned as a result of predict() call.}

\item{datasetname}{is the current datasetname using which we want to make predictions.}

\item{showCofusionMatrix}{logical, if TRUE the confusion matrix is generated (if it applies), if FALSE, confusion matrix is not generated.}
}
\value{
Displays the confusion matrix using the function confusionMatrix in the package caret
}
\description{
Creates a confusion matrix by cross-tabulating the observed and predicted classes with associated statistics.
}
