% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosstab with weights.R
\name{BSkyCrossTable}
\alias{BSkyCrossTable}
\title{Crosstab}
\usage{
BSkyCrossTable(
  data = NULL,
  x = NA,
  y = NA,
  layers = NA,
  weight = NA,
  digits = 3,
  max.width = 5,
  expected = FALSE,
  prop.r = FALSE,
  prop.c = FALSE,
  prop.t = FALSE,
  prop.chisq = FALSE,
  chisq = FALSE,
  fisher = FALSE,
  mcnemar = FALSE,
  resid = FALSE,
  sresid = FALSE,
  asresid = FALSE,
  missing.include = TRUE,
  dnn = NULL,
  datasetname = NULL,
  bSkyHandleSplit = TRUE
)
}
\arguments{
\item{x}{row variable}

\item{y}{column variable}

\item{layers}{one or more variables for layers}

\item{expected}{Expected counts are generated if this is TRUE}

\item{prop.r}{Row percentages are produced if this is TRUE}

\item{prop.c}{Column percentages are generated if this is TRUE}

\item{chisq}{if TRUE generates chi-square table}

\item{resid}{if TRUE, unstandardized residual are generated}

\item{sresid}{if TRUE, standardized residuals are generated}

\item{asresid}{if TRUE, adjusted residuals are generated}

\item{datasetname}{Name of the dataset from which x,y and layers (variables) are chosen}

\item{weights}{a numeric variable containing the frequency weights}
}
\value{
A list with the results
}
\description{
Creates crosstab with row, column and layer variables and optionally displays the following
Expected counts
Row and column percentages
Unstandardized, standardized and adjusted residuals
Chisq with odds ratio, McNemar and Fisher statistics
}
\examples{
BSky_Multiway_Cross_Tab = BSkyCrossTable(x=c('manufact'),y=c('model'),
layers=c('type'),datasetname='Dataset2',
chisq = FALSE,prop.r=FALSE,prop.c=FALSE,
resid=FALSE,sresid=FALSE,expected=FALSE,
asresid=FALSE)
BSkyFormat(BSky_Multiway_Cross_Tab)
}
