% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Factor Analysis.R
\name{BSkyFactorAnalysis}
\alias{BSkyFactorAnalysis}
\title{Factor Analysis}
\usage{
BSkyFactorAnalysis(
  vars,
  autoextraction = TRUE,
  factors = 1,
  rotation = "varimax",
  saveScores = FALSE,
  screeplot = FALSE,
  scores = "none",
  prefixForScores = "",
  dataset,
  excludeEnvPrefix = FALSE
)
}
\arguments{
\item{vars}{One or more numeric variables to extract factors from.}

\item{autoextraction}{Automatically determine the number factors or extract specific numbers of factors.}

\item{rotation}{determine the type of rotation and takes one of the values (none, quartimax, geominT, varimax, oblimin, simplimax, promax, geominQ and bentlerQ)}

\item{saveScores}{saves the factor scores in the dataset}

\item{screeplot}{If TRUE generates a screeplot.}

\item{dataset}{The dataset from which the 'vars' have been picked.}
}
\value{

}
\description{
Perform maximum-likelihood factor analysis on a covariance matrix or data matrix and generates a screeplot. Calls the function factanal in the stats package.
}
