% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSkyFactorVariableAnalysis.R
\name{BSkyFactorVariableAnalysis}
\alias{BSkyFactorVariableAnalysis}
\title{Factor Variable Analysis, top N}
\usage{
BSkyFactorVariableAnalysis(
  data = NULL,
  vars = NULL,
  show.only.top.factors = TRUE,
  max.number.top.factors = 30
)
}
\arguments{
\item{data}{The dataset name}

\item{vars}{The factor variables to display level counts}

\item{show.only.top.factors}{A boolean determining whether counts of only top levels or all levels should be displayed.}

\item{max.number.top.factors}{if show.only.top.factors = TRUE , specifies the number of levels to display counts for}
}
\value{
Factor variable analysis
}
\description{
Displays the counts of every level of each factor variable.
}
