% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSkyFormat.r
\name{BSkyFormat}
\alias{BSkyFormat}
\title{BSkyFormat generates formatted tabular data for the following R objects}
\usage{
BSkyFormat(obj, bSkyFormatAppRequest = FALSE, bSkyReturnObj = TRUE,
  ftable_change_variable_order = TRUE, sublist_length = 3,
  remove_rows_with_zero_count = FALSE, no_row_column_headers = FALSE,
  decimalDigitsRounding = BSkyGetDecimalDigitSetting(),
  engNotationSetting = BSkyGetEngNotationSetting(),
  singleTableOutputHeader = "")
}
\arguments{
\item{obj}{is an object to be formatted. This object must be one of the matrix, array, data.frame etc..}

\item{decimalDigitsRounding}{number of digits, the numeric data is rounded to. For scientific notatrion, prefix of the 'e' notaion is rounded.}

\item{engNotationSetting}{to display numbers as a scientific notation if TRUE}

\item{singleTableOutputHeader}{table title for a single table (no effect for the multiple table titles).

Parameters bSkyFormatAppRequest, bSkyReturnObj, ftable_change_variable_order, sublist_length, remove_rows_with_zero_count, no_row_column_headers are for internal use.}
}
\description{
matrix, array, data.frame, table, ftable, xtabs, htest, Glm, by, summary.glm, 
numSummary, tapply, summary.lm. summary.multinorm, multinorm, polr, summary.polr, vectors
List, Lm, Anova
To use BSkyFormat you must run it from within the BlueSky Statistics application.
}
\examples{
df <-data.frame(A=c(1,2,3), B=c(4,5,6), C=c(6,7,8))
BSkyFormat(df)
BSkyFormat(df, singleTableOutputHeader='My DataFrame Title') # to create a custom title for the formatted grid.
}

