% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSkyFrequency function.R
\name{BSkyFrequency}
\alias{BSkyFrequency}
\title{Frequency table}
\usage{
BSkyFrequency(data = NULL, vars = NULL)
}
\arguments{
\item{data}{The dataset name}

\item{vars}{selected variables}
}
\value{
frequency table
}
\description{
Generates the frequencies for every unique value in one or more variables or column names selected.
}
