% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetAllModelObj.r
\name{BSkyGetAvailableModels}
\alias{BSkyGetAvailableModels}
\title{Fetches currently available models}
\usage{
BSkyGetAvailableModels(
  objclasslist = c("lm", "glm", "randomForest"),
  returnClassTrain = TRUE
)
}
\arguments{
\item{objclasslist}{class of model to fetch}

\item{returnClassTrain}{}
}
\value{

}
\description{
Fetches currently available models of a specified classes.
}
