% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadDatasetFromPackage.r
\name{BSkyGetDatasetNameFromPackageDatasetList}
\alias{BSkyGetDatasetNameFromPackageDatasetList}
\title{Get dataset name from package dataset list}
\usage{
BSkyGetDatasetNameFromPackageDatasetList(datasetAndPackagedetails)
}
\value{

}
\description{

}
