% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Heatmap.R
\name{BSkyHeatMap}
\alias{BSkyHeatMap}
\title{HeatMap}
\usage{
BSkyHeatMap(
  rowVariable,
  colVariables,
  main = "Heatmap",
  Colv = FALSE,
  Rowv = TRUE,
  dendrogram = "row",
  scale = "none",
  noteForCells = FALSE,
  notecol = "cyan",
  density.info = c("none"),
  trace = "none",
  na.rm = TRUE,
  dataset,
  excludeEnvPrefix = FALSE
)
}
\arguments{
\item{rowVariable}{The column variable that contains information about the subjects that the graph is constructed for, see above}

\item{colVariables}{a list of column variables that contain values that will be represented as colors in the heatmap}

\item{Colv}{determines if and how the column dendrogram should be reordered. Has the options as the Rowv argument above and additionally when x is a square matrix, Colv="Rowv" means that columns should be treated identically to the rows.}

\item{Rowv}{determines if and how the row dendrogram should be reordered. By default, it is TRUE, which implies dendrogram is computed and reordered based on row means. If NULL or FALSE, then no dendrogram is computed and no reordering is done. If a dendrogram, then it is used "as-is", i.e. without any reordering. If a vector of integers, then dendrogram is computed and reordered based on the order of the vector.}

\item{dendrogram}{character string indicating whether to draw 'none', 'row', 'column' or 'both' dendrograms. Defaults to 'both'. However, if Rowv (or Colv) is FALSE or NULL and dendrogram is 'both', then a warning is issued and Rowv (or Colv) arguments are honored.}

\item{scale}{character indicating if the values should be centered and scaled in either the row direction or the column direction, or none. The default is "none".}

\item{noteForCells}{logical indicating whether values should be placed in the cells of the heatmap.}

\item{notecol}{color of notes. If specified values are noted in the heatmap cells in color specified.}

\item{density.info}{character string indicating whether to superimpose a 'histogram', a 'density' plot, or no plot ('none') on the color-key.}

\item{trace}{character indicating if the values should be centered and scaled in either the row direction or the column direction, or none. The default is "none".}

\item{na.rm}{logical determining whether NA’s should be removed}
}
\value{

}
\description{
A heat map is a graphical representation of data where the individual values contained in a matrix are represented as colors.  Facets can be optionally created by specifying a factor variable. You can also optionally specify themes, and specify a title and labels for the x and y axis
 The dataset must be in the format  column1, column2…columnN where
column1: Identifies the subjects e.g. names of people whose the cholesterol values are plotted
column2:  Identifies the cholesterol values at time 1
column3:  Identifies the cholesterol values at time 2
column4:  Identifies the cholesterol values at time 3
column5:  Identifies the cholesterol values at time 4…
Column1 is a label for each row in the heatmap.
For Column 2 onwards, values in each row of the column are represented as colors in the heatmap.
}
