% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HierarchicalCluster.R
\name{BSkyHierClus}
\alias{BSkyHierClus}
\title{Hierarchical Cluster}
\usage{
BSkyHierClus(
  varsToCluster,
  method,
  noOfClusters,
  distance,
  plotDendogram,
  assignClusterToDataset,
  label = "",
  plotBiplot,
  dataset,
  excludeEnvPrefix = FALSE
)
}
\arguments{
\item{varsToCluster}{The variables to analyze}

\item{method}{the agglomeration method to be used. This should be (an unambiguous abbreviation of) one of "ward.D", "ward.D2", 
"single", "complete", "average" (= UPGMA), "mcquitty" (= WPGMA),"median" (= WPGMC) or "centroid" (= UPGMC).}

\item{noOfClusters}{The number of clusters desired}

\item{plotDendogram}{Plot a dendogram True or false}

\item{assignClusterToDataset}{Save the cluster assignments to the dataset}

\item{label}{name for the new variable that stores the cluster assignments}

\item{plotBiplot}{plot Biplot TRUE or FALSE}

\item{dataset}{Name of the dataset}
}
\value{

}
\description{
Hierarchical cluster analysis on a set of dissimilarities and methods for analyzing it.
}
