% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndSmttest.R
\name{BSkyIndSmTTest}
\alias{BSkyIndSmTTest}
\title{t-test, independent samples}
\usage{
BSkyIndSmTTest(
  data = NULL,
  varNamesOrVarGlobalIndices = NULL,
  group = NULL,
  conf.level = 0.95,
  alternative = "two.sided",
  datasetNameOrDatasetGlobalIndex = NULL,
  missing = 0,
  bSkyHandleSplit = TRUE,
  excludeEnvPrefix = FALSE,
  cohens_d = FALSE,
  cohensd_correction = FALSE,
  hedges_g = FALSE,
  hedgesg_correction = FALSE,
  glass_d = FALSE,
  glassd_correction = FALSE
)
}
\arguments{
\item{varNamesOrVarGlobalIndices}{selected scale variables (say var1, var2)}

\item{group}{a factor variable with two levels (say var3)}

\item{conf.level}{a numeric value  (say 0.95).}

\item{alternative}{a character string specifying the alternative hypothesis, must be one of "two.sided"  (default), "greater" or "less". You can specify just the initial letter.}

\item{datasetNameOrDatasetGlobalIndex}{Name of the dataset (say Dataset)  from which var1, var2 and var3 are selected.}

\item{missing}{missing values are handled on a per variable basis (missing =0) or list wise across all variables (missing=1).}
}
\value{
A list with resulting tables is returned.
}
\description{
Performs a one sample t-tests against the two groups formed by a factor variable (with two levels). Displays results for equal variances TRUE and FALSE. For equal variances the pooled variance is used otherwise the Welch (or Satterthwaite) approximation to the degrees of freedom is used. Internally calls t.test in the stats package for every selected variable
}
\examples{
Dataset <- data.frame(Expenses=c(20,23,19,25,26), Sales=c(48,50,55,51,49), Gender=c('m','f','f','m','m'), Deptt=c('IT', 'Sales', 'IT','Sales','IT'))
BSky_One_Simple_T_Test = BSkyIndSmTTest(varNamesOrVarGlobalIndices =c('Sales','Expenses'),group=c('Deptt'),conf.level=0.95, alternative="less", datasetNameOrDatasetGlobalIndex ='Dataset')
}
