% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadRefreshDataFrame.r
\name{BSkyLoadRefresh}
\alias{BSkyLoadRefresh}
\title{Loading a dataframe in the BlueSky Statistics UI datagrid}
\usage{
BSkyLoadRefresh(
  bskyDatasetName,
  load.dataframe = TRUE,
  isRmarkdownOutputOn = BSkyIsRmarkdownOutputOn()
)
}
\arguments{
\item{bskyDatasetName}{is the dataframe name (use quotes around the name)}

\item{load.dataframe}{is provided so that you can selectively refresh or not refresh the UI datagrid. 
This parameter is defined in the dialog to make the decision, whether to refresh the UI datagrid or not. Default value is TRUE.}
}
\description{
loads the specified dataframe in the BlueSky Statistics UI datagrid.
}
\examples{
df <-data.frame(A=c(1,2,3), B=c(4,5,6), C=c(6,7,8))
BSkyLoadRefresh('df')
}
