% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataFrame.r
\name{BSkyLoadRefreshDataframe}
\alias{BSkyLoadRefreshDataframe}
\title{BSkyLoadRefreshDataframe loads the specified dataframe in the BlueSky Statistics UI datagrid. 
If dataframe is already loaded in the UI datagrid then it refresh it for any changes done.}
\usage{
BSkyLoadRefreshDataframe(dframe, load.dataframe = TRUE)
}
\arguments{
\item{dframe}{is the dataframe object(do not use quotes around the object)}

\item{load.dataframe}{is provided so that you can selectively refresh or not refresh the UI datagrid. 
This parameter is defined in the dialog to make the decision, whether to refresh the UI datagrid or not.
Default value is TRUE.}
}
\description{
To use BSkyLoadRefreshDataframe you must run it from within the BlueSky Statistics application.
}
\examples{
df <-data.frame(A=c(1,2,3), B=c(4,5,6), C=c(6,7,8))
BSkyLoadRefreshDataframe(df)
}

