% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_properties.r
\name{BSkyMakeMultiColumnFactor}
\alias{BSkyMakeMultiColumnFactor}
\title{Make Factor Variable}
\usage{
BSkyMakeMultiColumnFactor(colNames, dataSetNameOrIndex)
}
\arguments{
\item{colNames}{Variables to be converted to factor}

\item{dataSetNameOrIndex}{the dataset name that contains the columns those will be converted to factor}
}
\value{

}
\description{
The function factor is used to encode a vector as a factor (the terms ‘category’ and ‘enumerated type’ are also used for factors). If argument ordered is TRUE, the factor levels are assumed to be ordered. For compatibility with S there is also a function ordered.
is.factor, is.ordered, as.factor and as.ordered are the membership and coercion functions for these classes.
}
