% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSkyRandomForest.r
\name{BSkyMultiRandomForest}
\alias{BSkyMultiRandomForest}
\title{Optimal No of Trees}
\usage{
BSkyMultiRandomForest(x, y = NULL, startval, endval, stepval, mtry)
}
\arguments{
\item{x}{a data frame or a matrix of predictors, or a formula describing the model to be fitted (for the print method, an randomForest object).}

\item{y}{A response vector. If a factor, classification is assumed, otherwise regression is assumed. If omitted, randomForest will run in unsupervised mode.}

\item{startval}{starting value for ntree argument in randomForest().}

\item{endval}{ending value for ntree argument in randomForest().}

\item{stepval}{count by which startval should be incremented in every iteration till it reaches endval value.}

\item{mtry}{Number of variables randomly sampled as candidates at each split. Note that the default values are different for classification (sqrt(p) where p is number of variables in x) and regression (p/3)}
}
\value{
Prints the optimal number of trees. Also displays a table with ntree and the corresponding Out-of-Bag error
}
\description{
Starting with the default value of mtry, search for the optimal value (with respect to Out-of-Bag error estimate) of ntree (no of trees to grow) for randomForest. The function BSkyMultiRandomForest calls the function randomForest with the arguments defined below
}
