% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new-bskyonesm.r
\name{BSkyOneSmTTest}
\alias{BSkyOneSmTTest}
\title{t-test, one sample}
\usage{
BSkyOneSmTTest(
  data = NULL,
  varNamesOrVarGlobalIndices = NULL,
  mu = 0,
  conf.level = 0.95,
  alternative = "two.sided",
  datasetNameOrDatasetGlobalIndex = NULL,
  missing = 0,
  bSkyHandleSplit = TRUE,
  cohens_d = FALSE,
  cohensd_correction = FALSE,
  hedges_g = FALSE,
  hedgesg_correction = FALSE,
  glass_d = FALSE,
  glassd_correction = FALSE,
  debug = FALSE
)
}
\arguments{
\item{varNamesOrVarGlobalIndices}{selected scale variables (say var1, var2)}

\item{mu}{a number indicating the true value of the mean. (say 10)}

\item{conf.level}{a numeric value  (say 0.95).}

\item{alternative}{a character string specifying the alternative hypothesis, must be one of "two.sided"  (default), "greater" or "less". You can specify just the initial letter.}

\item{datasetNameOrDatasetGlobalIndex}{Name of the dataset (say Dataset)  from which var1, var2 and var3 are selected.}

\item{missing}{missing values are handled on a per variable basis (missing =0) or list wise across all variables (missing=1).}
}
\value{
A list with the results of the test
}
\description{
Performs one sample t-tests on selected variables. Optionally computes effect size indices for standardized differences: Cohen's d and Hedges' g (This function returns the population estimate.)
}
\examples{
Dataset <- data.frame(Expenses=c(20,23,19,25,26), Sales=c(48,50,55,51,49), Gender=c('m','f','f','m','m'), Deptt=c('IT', 'Sales', 'IT','Sales','IT'))
BSky_One_Simple_T_Test = BSkyOneSmTTest(varNamesOrVarGlobalIndices =c('Sales','Expenses'),mu=.9,conf.level=0.95,alternative ="two.sided", datasetNameOrDatasetGlobalIndex ='Dataset',missing=1)
}
