% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSkyCorrelation.R
\name{BSkyPlotCorrelationMatrix}
\alias{BSkyPlotCorrelationMatrix}
\title{Correlation Analysis}
\usage{
BSkyPlotCorrelationMatrix(
  data,
  vars = NULL,
  correlationType = "Pearson",
  missingValues = "complete.obs",
  visualizeCorrelation = TRUE,
  plotWeb = TRUE
)
}
\arguments{
\item{data}{a numeric matrix or data frame, or an object of class "rcorr.adjust" to be printed.}

\item{correlationType}{"pearson" or "spearman", depending upon the type of correlations desired; the default is "pearson".}

\item{missingValues}{how to handle missing data: "complete.obs", the default, use only complete cases; "pairwise.complete.obs", use all cases with valid data for each pair.}

\item{visualizeCorrelation}{prints PlotCorr()}

\item{plotWeb}{prints PlotWeb()}
}
\value{
Returns an object of class "rcorr.adjust", which is normally just printed.
}
\description{
This function uses the rcorr function in the Hmisc package to compute matrices of Pearson or Spearman correlations along with the pair wise p-values among the correlations. The p-values are corrected for multiple inference using Holm's method (see p.adjust). Observations are filtered for missing data, and only complete observations are used.
}
