% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries.R
\name{BSkyPlotTimeSeries}
\alias{BSkyPlotTimeSeries}
\title{Plot Time Series}
\usage{
BSkyPlotTimeSeries(
  vars,
  start,
  frequency,
  plot.type = "multiple",
  naturalLogYaxis = FALSE,
  main = "",
  ylab = "",
  dataset,
  excludeEnvPrefix = FALSE
)

BSkyPlotTimeSeries(
  vars,
  start,
  frequency,
  plot.type = "multiple",
  naturalLogYaxis = FALSE,
  main = "",
  ylab = "",
  dataset,
  excludeEnvPrefix = FALSE
)
}
\arguments{
\item{vars}{selected variables to plot}

\item{start}{Time of first observation should be entered in the format year,month or year,quarter e.g.( if your data is organized in months the 1992,1 for Jan 1992 or if your data is organized in quarters then 1992,1 refers to the first quarter of 1992.}

\item{frequency}{Number of observations in unit time. Example: for monthly there are 12 observation in a year. For quarterly there are 4 observation in a year.}

\item{plot.type}{"multiple" for separate and "single" for combined plot.}

\item{naturalLogYaxis}{if TRUE an Y axis is shown as natural log value.}

\item{main}{main title of the plot}

\item{ylab}{title for the y axis}

\item{dataset}{the name of the dataset from which the vars has been picked.}
}
\value{



}
\description{
Creates time series plot in combined or separately. When combined multiple variables are plotted together, when separate each variable is plotted independently. BSkyPlotSeriesWithCorrelations is a wrapper function that creates a time series object and internally calls plot.ts.

Creates time series plot in combined or separately. When combined multiple variables are plotted together, when separate each variable is plotted independently. BSkyPlotSeriesWithCorrelations is a wrapper function that creates a time series object and internally calls plot.ts.
}
