% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.r
\name{BSkyPredict}
\alias{BSkyPredict}
\title{Score a dataset using a model}
\usage{
BSkyPredict(
  modelname = "multinom",
  prefix = "multinom",
  confinterval = FALSE,
  level = 0.95,
  datasetname = "Dataset4"
)
}
\arguments{
\item{modelname}{a model object for which prediction is desired.}

\item{prefix}{prefix string that will be used to create new variables containing the predictions.}

\item{datasetname}{is the current dataset to score and save predictions to.}
}
\value{

}
\description{
Model scoring does the following
1. Scores the current dataset using the selected prebuilt model. Stores predictions with the specified confidence interval in the current dataset using the specified prefix.
2. Optionally creates a confusion matrix and a ROC curve
predict is a generic function for making predictions using the selected model.
}
