% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Principan Component Analysis.R
\name{BSkyPrinCompAnalysis}
\alias{BSkyPrinCompAnalysis}
\title{Principal Component Analysis}
\usage{
BSkyPrinCompAnalysis(
  vars,
  cor = FALSE,
  componentsToRetain = 0,
  generateScreeplot = FALSE,
  prefixForComponents = "",
  dataset,
  excludeEnvPrefix = FALSE
)
}
\arguments{
\item{vars:}{The variables in a character vector to extract components from}

\item{cor:}{A boolean that specifies whether the calculation should use a correlation or covariance matrix}

\item{componentsToRetain:}{A numeric that Specifies the number of components to retain in the dataset. A new variable is created in the dataset for each component invoked}

\item{​generateScreeplot:}{Generates a screeplot}

\item{prefixForComponents:}{Prefix to use when saving the components to a dataset}

\item{dataset:}{The name of the dataset as a string}
}
\value{
princomp returns a list with class "princomp" containing the following components
sdev: the standard deviations of the principal components.
loadings: the matrix of variable loadings (i.e., a matrix whose columns contain the eigenvectors). This is of class "loadings": see loadings for its print method.
center: the means that were subtracted.
scale: the scalings applied to each variable.
n.obs: the number of observations.
scores: if scores = TRUE, the scores of the supplied data on the principal components. These are non-null only if x was supplied, and if covmat was also supplied if it was a covariance list. For the formula method, napredict() is applied to handle the treatment of values omitted by the na.action.
call: the matched call.
na.action: If relevant.
}
\description{
Performs a principal components analysis on the given numeric data matrix and returns the results as an object of class princomp. Internally calls princomp in the stats package.
}
