% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fixed and Random effects.r
\name{BSkyprint.summary.merMod}
\alias{BSkyprint.summary.merMod}
\title{Mixed models}
\usage{
BSkyprint.summary.merMod(
  x,
  digits = max(3, getOption("digits") - 3),
  correlation = NULL,
  symbolic.cor = FALSE,
  signif.stars = getOption("show.signif.stars"),
  ranef.comp = c("Variance", "Std.Dev."),
  show.resids = TRUE,
  ...
)
}
\arguments{
\item{x}{}

\item{digits}{}

\item{correlation}{}

\item{symbolic.cor}{}

\item{signif.stars}{}

\item{ranef.comp}{}

\item{show.resids}{}
}
\value{

}
\description{
Print model summary
}
