% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NonAnalytics.R
\name{BSkyRecode}
\alias{BSkyRecode}
\title{Recode variables}
\usage{
BSkyRecode(
  colNames,
  newColNames = "",
  OldNewVals,
  NewCol = FALSE,
  prefixOrSuffix = "",
  prefixOrSuffixString = "",
  makeNumeric = FALSE,
  dontMakeFactor = FALSE,
  dataSetNameOrIndex
)
}
\arguments{
\item{colNames}{A character vector containing one or more variables in the dataset to recode}

\item{newColNames}{A character vector containing the names of the new columns.}

\item{OldNewVals}{A character string of recode specifications in the form oldval1,newval1, oldval2,newval2}

\item{NewCol}{A Boolean indicating whether recoded values are stored in new  variables (TRUE) or existing variables are overwritten(FALSE).}

\item{prefixOrSuffix}{Specify if you want to store the recoded values in new variables prefixed or suffixed with the name you specify.  Enter prefix or suffix.}

\item{prefixOrSuffixString}{Enter a string to use as a prefix or suffix to the existing variable name. Recoded values will be stored in these variables.}

\item{dataSetNameOrIndex}{The dataset/dataframe name}
}
\value{

}
\description{
Recodes one or more a numeric vector, character vector, or factors according to  recode specifications. You can store the results by overwriting existing variables, specifying new variable names to store recoded values or choosing to store the recoded values in  new variables with a suitable prefix or suffix. the prefix or suffix will be applied to the existing variable name.
Note: We will not convert from numeric to factor. When  a numeric is recoded, it will remain a numeric, when a factor variable is recoded it will remain a factor.
}
