% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UA_UI_Functions.R
\name{BSkySetDataFrameSplit}
\alias{BSkySetDataFrameSplit}
\title{Set or remove split}
\usage{
BSkySetDataFrameSplit(col.names, datasetnameorindex, removeall.splits = FALSE)
}
\arguments{
\item{col.names}{These are the column names/variable names that you want to split the dataset by, e.g. col.names =c("var1", "var2"). To reset the split, set to c().}

\item{datasetnameorindex}{this is the name of the index.​}

\item{removeall.splits}{TRUE splits are removed, FALSE splits are added.}
}
\value{

}
\description{
Splits the data into groups based on the factors selected, once the dataset is split, the analysis you select is performed independently for each split. For example if you run a crosstabulation analysis or a hypothesis test, this analysis is performed independently for each split (the output of the analysis is also generated separately for each split). 
OR
Removes the split on the dataset. Splits, see Data>Group By>Set Split splits the data into groups based on the factors selected, once the dataset is split, the analysis you select is performed independently for each split. For example if you run a crosstabulation analysis or a hypothesis test, this analysis is performed independently for each split (the output of the analysis is also generated separately for each split).
}
