% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSkyFormat.r
\name{BSkySetHtmlStylingSetting}
\alias{BSkySetHtmlStylingSetting}
\title{Set HTML style}
\usage{
BSkySetHtmlStylingSetting(
  tableTheme = "kable_styling",
  fontFamily = "Helvetica",
  tableHeaderBackgroundColor = "#bdbdbd",
  ColumnAlign = "Right",
  tablePosition = "Left",
  fullWidthDisplay = TRUE,
  tableOuterBorder = FALSE,
  tableGridLines = FALSE,
  striped = TRUE,
  hover = TRUE,
  latex_scale_down = FALSE,
  latex_hold_position = FALSE,
  columHeaderScrollFixed = TRUE,
  fontSize = 0,
  overrideStylingSettings = TRUE
)
}
\value{

}
\description{

}
