% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSkySetWeight.r
\name{BSkySetWeight}
\alias{BSkySetWeight}
\title{Expand Weights}
\usage{
BSkySetWeight(weights, data, newdata)
}
\arguments{
\item{weights}{The dataset variable that contains the weights.}

\item{data}{The input data.frame or data.table.}

\item{newdata}{The new dataset where the rows are replicated for the weights specified.}
}
\value{
expanded dataset according to the weights
}
\description{
Creates a new dataset with rows expanded as per weights. Expands (replicates) the rows of a data.frame  by a value contained in one of the columns in the source data.frame or data.table. BSkySetWeights() calls expandRows() from the package splitstackshape.
}
