% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcmdr.functions.R
\name{BSkyStandardizeVars}
\alias{BSkyStandardizeVars}
\title{Standardize Variable(s)}
\usage{
BSkyStandardizeVars(
  vars,
  center,
  scale,
  stingToPrefixOrSuffix,
  prefixOrSuffix,
  datasetname,
  excludeEnvPrefix = FALSE
)
}
\arguments{
\item{vars:}{One or more variables to standardize. Only numeric variables (not factors) supported.}

\item{center:}{If center is TRUE then centering is done by subtracting the column means (omitting NAs) of x from their corresponding columns, and if center is FALSE, no centering is done.}

\item{scale:}{If scale is TRUE then scaling is done by dividing the (centered) columns of x by their standard deviations if center is TRUE, and the root mean square otherwise. If scale is FALSE, no scaling is done.}

\item{stingToPrefixOrSuffix:}{A character string that specifies the prefixor suffix  to use for the new standardized variables( i.e. new columns in the dataset).}

\item{prefixOrSuffix:}{specify if you want a prefix or a suffix}

\item{datasetname:}{The dataset/dataframe name}
}
\value{

}
\description{
Standardizes variables (z scores).  The standardized values are stored in new variables with either the prefix or suffix of the original variables. The option is provide to center and/or scale.
Note: We will not convert from numeric to factor. When  a numeric is recoded, it will remain a numeric, when a factor variable is recoded it will remain a factor.
}
