% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Table-print-function.R
\name{BSkySummaryStats}
\alias{BSkySummaryStats}
\title{Summary statistics}
\usage{
BSkySummaryStats(
  data = NULL,
  datasetColumnObjects = list(),
  groupByColumnObjects = list(),
  datasetName = c(),
  stats = c(min = TRUE, max = TRUE, mean = TRUE, median = TRUE, quantiles = TRUE),
  quantilesProbs = c(0, 0.25, 0.5, 0.75, 1),
  additionalStats = c(),
  maxsum = 30,
  bSkyFormatAppRequest = FALSE,
  ftable_change_variable_order = TRUE,
  sublist_length = 3,
  remove_rows_with_zero_count = FALSE,
  no_row_column_headers = FALSE
)
}
\arguments{
\item{datasetColumnObjects:}{selected scale variables (say Dataset$var1, Dataset$var2)}

\item{groupByColumnObjects:}{one or more factor variables to group by (say  Dataset$var3, Dataset$var4)}

\item{statFunctionList:}{List of functions. The ones set to TRUE will be executed. (say min=TRUE, sd=TRUE)​}

\item{​quantilesProbs:}{Probabilities of the quantiles}

\item{additionalStats:}{Addition statistical function that user can pass ( say var)}

\item{datasetName:}{Name of the dataset from which datasetColumnObjects and groupByColumnObjects are chosen}
}
\value{
An object of class "data.frame", giving the results for each function on each variable.
}
\description{
Outputs the  following numerical statistics:
min, max, mean, median, sum, sd, stderror, iqr,  quartiles. If quartiles is selected, you can specify the comma separated quartiles needed.
In addition to these, the user can pass, a list of comma separated statistical function names for example var.
}
\examples{
Dataset <- data.frame(Expenses=c(20,23,19,25,26), Sales=c(48,50,55,51,49), Gender=c('m','f','f','m','m'), Deptt=c('IT', 'Sales', 'IT','Sales','IT'))</br>
Result_Numerical_Statistics_Analysis = BSkySummaryStats(datasetColumnObjects = list(Sales = Dataset$Sales, Expenses = Dataset$Expenses), groupByColumnObjects = list(Deptt= Dataset$Deptt), stats = c(min=FALSE,max=FALSE,mean=TRUE,median=TRUE,sum=FALSE,sd=FALSE,stderror=FALSE,iqr=FALSE,quantiles=FALSE),datasetName="Dataset" )
}
