% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSkyVersion.r
\name{BSkyVersion}
\alias{BSkyVersion}
\title{BlueSky package adn other details}
\usage{
BSkyVersion()
}
\value{
prints details about system, R session etc.
}
\description{
Get BlueSky package details along with R session related details.
}
\examples{
BSkyVersion()
}
