% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-cgit.R
\docType{data}
\name{cgit}
\alias{cgit}
\alias{cgit_investments_df}
\alias{cgit_construction_contracts_df}
\alias{cgit_troubled_transactions_df}
\alias{cgit_combined_df}
\title{China Global Investment Tracker}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1635 rows and 12 columns.}
\source{
https://www.aei.org/china-global-investment-tracker/
}
\usage{
cgit_investments_df

cgit_construction_contracts_df

cgit_troubled_transactions_df

cgit_combined_df
}
\description{
\strong{The China Global Investment Tracker (CGIT) is the only comprehensive public data set covering China’s global investment and construction, which are documented both separately and together. Inaugurated in 2005, the CGIT now includes over 3,300 large transactions across energy, transportation, real estate, and other industries, as well as 300 troubled transactions. The full list, with the amount, Chinese parent company, host country, and sector, is available for public use with the proper citation. The tracker is published by the American Enterprise Institute.}
}
\examples{
cgit_investments_df

cgit_construction_contracts_df

cgit_troubled_transactions_df

cgit_combined_df
}
\references{
\itemize{
\item The American Enterprise Institute and The Heritage Foundation
}
}
\keyword{datasets}
