% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-cmcs.R
\docType{data}
\name{cmcs_licenses_df}
\alias{cmcs_licenses_df}
\alias{cmcs_special_areas_df}
\title{Computerized Mining Cadastre System (CMCS)}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2782 rows and 7 columns.}
\source{
\href{https://cmcs.mram.gov.mn/cmcs}{CMCS Online}
}
\usage{
cmcs_licenses_df

cmcs_special_areas_df
}
\description{
Data made available by the Mongolia's Computerized Mining Cadastre System (CMCS).
}
\examples{
head(cmcs_licenses_df)

head(cmcs_special_areas_df)

}
\author{
Brendan Knapp \email{brendan.g.knapp@nps.edu}
}
\keyword{datasets}
