% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-eitim.R
\docType{data}
\name{eitim_licenses_sf}
\alias{eitim_licenses_sf}
\alias{eitim_companies_df}
\title{Extractive Industries Transparency Initiative (EITI)}
\format{An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 3447 rows and 14 columns.}
\source{
\href{https://e-reporting.eitimongolia.mn}{EITIM E-Reporting System}
}
\usage{
eitim_licenses_sf

eitim_companies_df
}
\description{
Data made available by the Extractive Industries Transparency Initiative' Mongolia site.
}
\examples{
head(eitim_licenses_sf)

head(eitim_companies_df)

}
\author{
Brendan Knapp \email{brendan.g.knapp@nps.edu}
}
\keyword{datasets}
