% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwCSV.R
\name{UAwriteCSV}
\alias{UAwriteCSV}
\title{UAwriteExcel writes a in memory dataset to a disk file in Excel format}
\usage{
UAwriteCSV(csvfilename, dataSetNameOrIndex)
}
\arguments{
\item{csvfilename}{is a full path filename of a new CSV file that is to be created.}

\item{dataSetNameOrIndex}{is a name of a dataset that is a reference for loaded Excel dataset.}
}
\description{
UAwriteExcel writes a in memory dataset to a disk file in Excel format
}
