% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-wdi.R
\docType{data}
\name{world_bank_df}
\alias{world_bank_df}
\title{World Bank's World Development Indicators}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 7274562 rows and 6 columns.}
\source{
\href{http://datatopics.worldbank.org/world-development-indicators/}{World Bank}
}
\usage{
world_bank_df
}
\description{
\emph{The World Development Indicators is a compilation of relevant, high-quality, and internationally comparable statistics about global development and the fight against poverty. The database contains 1,600 time series indicators for 217 economies and more than 40 country groups, with data for many indicators going back more than 50 years.}
}
\examples{
world_bank_df

}
\author{
Brendan Knapp \email{brendan.g.knapp@nps.edu}
}
\keyword{datasets}
