\name{BlueSky-package}
\alias{BlueSky-package}
\alias{BlueSky}
\docType{package}
\title{
BlueSky Collection of Functions for Data Access, Analysis, and Formatting.
}
\description{
This package provides a variety of useful functions for accessing data, analyzing data and creating publication-quality output
in text, HTML, and LaTeX.
The functions are used extensively in the BlueSky Statistics user interface to R, but are useful to any R users.
}
\details{
\tabular{ll}{
Package: \tab BlueSky\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2021-10-15\cr
License: \tab GPL (version 3 or later)\cr
}
Package can be used with BlueSky Statistics application or any other R environment

}
\author{
BlueSky Statistics.

Maintainer: contact us <support@blueskystatistics.com>

}
\references{
~~ Literature or other references for background information ~~
}



