% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSkyFormat.r
\name{BSkyFormat}
\alias{BSkyFormat}
\title{BSkyFormat generates formatted output (text or table)}
\usage{
BSkyFormat(
  obj,
  maxOutputTables = BSkyGetTableDisplayLimits(),
  outputTableIndex = c(),
  outputColumnIndex = c(),
  outputTableRenames = c(),
  outputColumnRenames = c(),
  outputColumnRenamesRow = c(),
  maxRowLimit = BSkyGetTableDisplayLimits(),
  maxColLimit = BSkyGetTableDisplayLimits(),
  silentFormatting = FALSE,
  bSkyFormatAppRequest = FALSE,
  bSkyReturnObj = TRUE,
  ftable_change_variable_order = TRUE,
  sublist_length = 3,
  remove_rows_with_zero_count = FALSE,
  no_row_column_headers = FALSE,
  decimalDigitsRounding = BSkyGetDecimalDigitSetting(),
  engNotationSetting = BSkyGetEngNotationSetting(),
  singleTableOutputHeader = "",
  repeatAllTableFooter = c(),
  perTableFooter = c(),
  isRound = BSkyGetRound(),
  coefConfInt = 0.95,
  pvalueDisplaySettings = BSkyGetPvalueDisplaySetting(),
  isKableOutput = TRUE,
  isLatexOutput = FALSE,
  isRmarkdownOutput = TRUE,
  isTextOutput = FALSE,
  getNonRenderedTables = FALSE,
  ignoreEnvStyleOverride = FALSE,
  forceColumnAlign = c(),
  kableStyleTheme = "kable_styling",
  textTableFormat = BSkyGetTextTableFormat(),
 
    tableStylingOptions = "table_border = F, column_align = r, header_background = \\"#bdbdbd\\" , more_options = c(bootstrap_options = c(\\"striped\\", \\"hover\\", \\"condensed\\", \\"responsive\\"), position = \\"left\\", full_width = F, html_font = \\"Helvetica\\", fixed_thead = list(enabled = T, background = \\"#bdbdbd\\"))",
  RM.SSP = TRUE,
  RM.SSPE = RM.SSP
)
}
\arguments{
\item{obj}{is an object of type data.frame, matrix, list, text or BSkyReturnStructure object, that is to be formatted.}

\item{maxOutputTables}{maximum tables to output if obj contains mulitple tables}

\item{outputTableIndex}{the order in which output tables are printed}
}
\value{

}
\description{

}
